Add more descriptive features and factory bot

+++ db/migrate/20180925150330_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration[5.2]
     create_table :products do |t|
       t.string :name
       t.text :description
-      t.integer :price
+      t.money :price
 
       t.timestamps
     end
